# Redid structure of site
I reviewed the entire structure of the project and the design of the commits.
Here are the main points that have been changed

> Today **cases** are named **projects**.

## Page Structure

All pages now start with `index.html` and `project.css`.
No need to redefine the base files.
The directory structure is consistent across all project.

## Projects Structure

- Project are now categorized as:
  - `projects`
  - `personal-projects`
  - `other-pages`

## CSS Structure Improvements

- The CSS has been restructured into two modules:
  - `base.css`
  - `projects.css`
- This modules have a set items:
  - `base-items`
  - `projects-items`
- There are independent items for grids like:
  - `link-grid.css`
  - `poster-grid.css`
  - `projects-grid.css`

## Theme Addition

Add the **dark/light theme**. Made a `colors.css` in `base.css` module.
All pages adopted for it. Redid all `projects.css` item for it.

## Resource Management

Moved all download resources to a hidden folder to keep GitHub clean _it not pushed_.

## Design and Adaptability Fixes

Fixes most of the aspects about design and adaptability

## Added, fixed and finished

* [ ] add personal
* [ ] add posters
* [ ] add stickers
* [ ] add all-pages _for Alpha_
- [x] finish resume
- [x] finish melectron
- [x] finish morfin
- [x] finish desktops
- [x] add&finish showreel
- [x] add&finish useful
- [x] add&finish 1523tv
- [x] add&finish birthday
- [x] add&finish kfbrand
- [x] add&finish minecraft-posters
- [x] add&finish morphinism
- [x] add&finish morphinism-lyrics
- [x] add&finish nuqun
- [x] add&finish wild-matching
+ add fonts for projects
+ add icons pop-up
+ optimized all scripts

## For future

- Also, invisible changes are ready for future projects, in the TEST folder _it not pushed_.